KSPM: Fix behavior when a selection is made while resetting.
The KSPM should never clear or reset the selection of the
QItemSelectionModel it refers to.  It should be a read-only view on
that selection.  If the selection changes while the source is being
reset, ignore that selection change.  The proxy will clear its own
cached data when it processes the modelReset signal from the model,
and will populate itself with the selection as it is at that point.