Fix build with Qt 5.5
Turns out that there's apparently no overload accepting a QPointer as
the first argument.

This is on Gentoo with qtbase e374ffc29c67493a51527117b55a53dfa5dd4267
and GCC 4.8.3.

Also make sure to tie the lifetime of the connection with `this` because
the lambda accesses our D-pointer, which means that it has to be
disconnected on destruction of `this`. Thanks to Milian Wolff for
catching this.

REVIEW: 123679