KSPM: Simplify row removal handling
The KSelectionProxyModelPrivate::beginRemoveRows method is responsible
for figuring out the range in the proxy which is to be removed as a
result of removal in the source model.

There are some trivial cases which were made overcomplicated by the old
code.  The complex bit is figuring out the interaction with the
top-level.  I suspect that can be simplified further.