KRecursiveFilterProxyModel: fix QSFPM corruption due to filtering out rowsRemoved signal
If the row being removed from the source model isn't shown by KRFPM,
it's correct that we don't need to emit any signals, however we still
need to tell our base class QSFPM so that it updates the mappings
of the visible siblings below that removed row.

Testcase: deleting the Last Search folder in kmail
(full integration test for this testcase will be added to mailcommon).

REVIEW: 128428
BUG: 349789
FIXED-IN: 5.25