KExtraColumnsProxyModel: Persist model indexes after emitting layoutChange, not before
Same fix as Stephen Kelly's fix for QIdentityProxyModel in https://codereview.qt-project.org/180390

I tried to write a unittest but QStandardItemModel doesn't support moveRow,
and we don't have all the infrastructure here to force emitting a layoutChange
with specific parents like in the Qt unittests.

This commit, added to the 3 fixes in Qt, fixes crashes when moving folders in kmail.

CCMAIL: steveire@gmail.com