Add KColumnHeadersProxyModel
Summary:
It converts a model's column headers into a simple list model. It is
mainly useful as a helper to create headers for QtQuick TableView but
may also be useful for other things.

Test Plan: The added test QML file shows a single 1 as described in the test.

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: kossebau, vkrause, davidedmundson, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25311