Fix binding to temporary object due to auto-converted loop variables
There is no reason for the loop variables to be of type
QPersistentModelIndex, as they are used and copied to the class member
as QModelIndex type in the loop body only.

So changing the reference type to match the looped list element type
also fixes the current resulting binding to a temporary object.

NO_CHANGELOG