Fix assertions in KDescendantsProxyModel
Remove assertions in `kdescendantsproxymodel.cpp` that will make the program crash when the model is empty.

Note: if `hasChildren` returns true, it doesn't mean that `rowCount` for the same index must be >0.
For example, KDirModel's `hasChildren` returns true for directories with >=1 children, but its `rowCount` doesn't list
the children and just returns 0 if they weren't already listed.

BUG: 452043