KSortFilterProxyModel: Don't recurse into syncRoleNames
Row count of a proxy model changes as it filters its source model, but
that's not a good reason to refresh role names every time. Instead we
are only interested in source model's changes.

This reduces amount of insert/remove back and forth signals during
filtering, and fixes the permanent "ghost delegate" at the end of a
ListView.

CCBUG: 476950