KSortFilterProxyModel: Fix multiple sources of truth confusion
Sort and filter roles may come from upstream model's properties (by ID)
or from our class (by role name), and it is important to keep track
which one was set explicitly and thus should be considered the source
of truth at any given moment in time -- especially if/when source model
is set afterwards such that these pairs of properties can not sync
immediately and have to be delayed instead.

To solve this problem, the an object now stores a flag which determined
the source of truth per each pair, and another flag to prevent
recursion and switch to "Role ID" source when *RoleChanged signals are
emitted NOT as a side-effect of us updating a corresponding *RoleName.

BUG: 476950